backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
PyJWT==2.8.0psycopg2-binary==2.9.9
requests==2.31.0
